requirement.txt
flake8
pytest
pyyaml
requests
requests-mock
mock
